Add GoogleTest Rust to the project list
I wasn't entirely sure whether to put this under "Frameworks and Runners" or "Mocking and Test Data". I put it under the former due to its support for non-fatal assertions, which pushes it a bit more in the direction of a test framework. But I'm not picky about that.

As per guidance:
 * https://github.com/google/googletest-rust has currently 151 stars
 * https://crates.io/crates/googletest currently shows 17,693 downloads